chore: move utils functions out of main.go
